+++ lib/src/constants.dart
@@ -23,6 +23,24 @@ class BnApiUrls {
   static const String optionsApiVersion = 'v1';
 }
 
+class BnApiOrderType {
+  static const String orderTypeLimit = 'LIMIT';
+  static const String orderTypeMarket = 'MARKET';
+  static const String orderTypeStopLoss = 'STOP_LOSS';
+  static const String orderTypeStopLossLimit = 'STOP_LOSS_LIMIT';
+  static const String orderTypeTakeProfit = 'TAKE_PROFIT';
+  static const String orderTypeTakeProfitLimit = 'TAKE_PROFIT_LIMIT';
+  static const String orderTypeLimitMaker = 'LIMIT_MAKER';
+
+  static const String futureOrderTypeLimit = 'LIMIT';
+  static const String futureOrderTypeMarket = 'MARKET';
+  static const String futureOrderTypeStop = 'STOP';
+  static const String futureOrderTypeStopMARKET = 'STOP_MARKET';
+  static const String futureOrderTypeTakeProfit = 'TAKE_PROFIT';
+  static const String futureOrderTypeTakeProfitMarket = 'TAKE_PROFIT_MARKET';
+  static const String futureOrderTypeLimitMaker = 'LIMIT_MAKER';
+}
+
 class BnAggKeys {
   // For accessing the data returned by Client.aggregate_trades().
   static const String aggID = 'a';
@@ -33,4 +51,10 @@ class BnAggKeys {
   static const String aggTime = 'T';
   static const String aggBuyerMakers = 'm';
   static const String aggBestMatch = 'M';
+}
+
+class BnTimeInForce {
+  static const String timeInForceGtc = 'GTC'; // Good till cancelled
+  static const String timeInForceIoc = 'IOC'; // Immediate or cancel
+  static const String timeInForceFok = 'FOK'; // Fill or kill
 }