+++ lib/src/objects.dart
@@ -1,8 +1,8 @@
 enum KLinesType { spot, futures, futuresCoin }
+
 enum HttpMethod { get, post, put, delete }
 
 enum Status { preTrading, trading, postTrading, endOfDay, halt, auctionMatch, breakTrading }
-
 
 /// Maps [Status] to a Binance string
 const statusMap = <String, Status>{
@@ -15,7 +15,6 @@ const statusMap = <String, Status>{
   'BREAK': Status.breakTrading
 };
 
-
 class BinanceApiException implements Exception {
   final int code;
   final String message;
@@ -25,7 +24,6 @@ class BinanceApiException implements Exception {
   @override
   String toString() => message;
 }
-
 
 class SymbolProduct {
   final String symbol;
@@ -38,17 +36,17 @@ class SymbolProduct {
   final String quoteAsset;
   final num quotePrecision;
 
-  final List<OrderType> orderTypes;
+  // final List<OrderType> orderTypes;
   final bool icebergAllowed;
   // List<Filter> filters;
 
-  Symbol.fromMap(Map m)
+  SymbolProduct.fromMap(Map m)
       : symbol = m['symbol'],
         status = statusMap[m['status']]!,
         baseAsset = m['baseAsset'],
         baseAssetPrecision = m['baseAssetPrecision'],
         quoteAsset = m['quoteAsset'],
         quotePrecision = m['quotePrecision'],
-        orderTypes = List<String>.from(m['orderTypes']).map((s) => orderTypeMap[s]!).toList(),
+        // orderTypes = List<String>.from(m['orderTypes']).map((s) => orderTypeMap[s]!).toList(),
         icebergAllowed = m['icebergAllowed'];
 }