+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 
-| Version |    Date    | Changes    |
-|:-------:|:----------:|:-----------|
-|  0.0.1  | 15.10.2022 | BaseClient |
+| Version |    Date    | Changes                                              |
+|:-------:|:----------:|:-----------------------------------------------------|
+|         |            | Tests                                                |
+|         |            | Comments                                             |
+|         |            | Not serialized data for compute                      |
+|         |            | Serialized data                                      |
+| 0.0.10  |            | C2C Endpoints                                        |
+| 0.0.10  |            | Fiat Endpoints                                       |
+| 0.0.9a  |            | Options API: Account and trading interface endpoints |
+|  0.0.9  |            | Options API: Quoting interface endpoints             |
+|  0.0.8  |            | COIN Futures API                                     |
+|  0.0.7  |            | Futures API                                          |
+|  0.0.6  | 17.10.2022 | Sub Accounts                                         |
+|  0.0.6  | 17.10.2022 | Lending Endpoints                                    |
+|  0.0.6  | 17.10.2022 | Margin OCO, Cross-margin, Isolated margin stream     |
+|  0.0.5  | 16.10.2022 | Margin Trading Endpoints                             |
+|  0.0.4  | 16.10.2022 | Withdraw Endpoints                                   |
+|  0.0.3  | 16.10.2022 | User Stream Endpoints                                |
+|  0.0.2  | 16.10.2022 | Order API                                            |
+|  0.0.1  | 15.10.2022 | BaseClient                                           |